Remove use and rename package to gatekeeper
